remove unused order_count parameter
It seems that in dillo-0.8.0-css-3 it was used to make sure that later
definitions of the same CSS property have more weight than previous
ones. If for example a first stylesheet sets background-color to
black, and a second stylesheet sets it to green, green should win.
But as we currently parese everything in order (throughing away all
parsed CSS data when a new CSS stylesheet has arrived), we don't
need to remember the original ordering of stylesheets.