fix image scaling on reload with border, margin, or padding > 0
On reload images have their buffer set already from cache when
Image::sizeAllocateImpl() is called.
In that case the widget is still at size 1x1. In that case
 getContentHeight() and getContentWidth() return values <= 0
when border + margin + padding > 0.

To fix the issue use the new size given in allocation instead of the
current widget dimensions to determine whether to scale the buffer.

(report and initial analysis by furaisanjin <furaisanjin@gmail.com>)