check scrollbar visibility before sending it events
There's probably useful untangling of the scrollbar code that
one could do because they're probably supposed to resize to
zero thickness when we can't see them, but...

And as for the modifier check, it was originally added in
http://cvs.auriga.wearlab.de/cgi-bin/cvsweb.cgi/dw2/dw/fltkviewport.cc?cvsroot=dillo#rev1.21
which just says "Fixed an include for OpenBSD compatibility." --
so I have no idea whether it's there for a reason. If a reason
reveals itself, I can add a comment.