Second patch attempt at the FL::fl_oldfocus problem.
Since fltk-1.3.3, FL::fl_oldfocus became private. This patch allows linking
dillo with fltk-1.3.3.

The previous patch had this problem:
"going to a page that has a Fl_Input, give focus to the input, move the
cursor out of the dillo window, and move it back into the dillo window,
focus does not return to the input."

This one solves both problems.